use more fp code
