+++ apps/nextjs/src/lib/analytics/keyStages.test.ts
@@ -1,4 +1,4 @@
-import { parseKeyStage } from "./helpers";
+import { parseKeyStage } from "./keyStages";
 
 const testKeyStageMap: Record<
   string,
